Factor out mask generation to objective code
